requirements.txt
openai>=1.12.0
elevenlabs>=0.2.24
requests>=2.31.0
moviepy>=1.0.3
python-dotenv>=1.0.0
Pillow>=10.0.0
pydub>=0.25.1
pexels-api>=1.0.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.80.0
TikTokApi>=5.3.3
praw>=7.7.0
schedule>=1.1.0
pytube>=15.0.0
feedparser>=6.0.10
flask>=3.0.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
scipy>=1.11.0
selenium>=4.15.0
playwright>=1.40.0
fake-useragent>=1.4.0
google-cloud-texttospeech>=2.14.0
schedule>=1.2.0
flask>=3.0.0
